Moves all the tf stuff to separate module
