+++ examples/demo14.yaml
@@ -1,3 +1,5 @@
+modules:
+    - galsim.tensorflow
 
 psf :
     type : Gaussian
