+++ galsim/tensorflow/__init__.py
@@ -1,0 +1,2 @@
+from .input_generative import *
+from .generative_model import *
