+++ galsim/tensorflow/generative_model.py
